Added local storage usage to simplify navigation between components and remove passing of user session related parameters (username, email, token...). Now the only parameters passed by query on router push are the non-user related (like house_id for HouseDetails or MyHouseDetails).
The way this is done is by using method persist on Login and Register (this saves parameters on localStorage) and then loading them on every page that needs them with mounted funtion (similar to created).

Finally, also added 'My houses' section to bottom of profile page with horizontal navigacion carousel (added new component MyHouseDetails to show details of this specific houses so that they don't display dates or submit button).